Cleaned up logs, split cookie functionality into its own file, made sure tests were all passing, added global 'canSsr' to hold reference to the global document and request object.
