+++ .gitignore
@@ -1,5 +1,3 @@
-.DS_Store
-
 node_modules/
 test/tests/node_modules/
 doc/
