+++ lib/make_wait_options.js
@@ -1,6 +1,6 @@
 var Override = require("can-wait").Override;
 
-module.exports = function(doc){
+module.exports = function ( doc, req ) {
 	var overrides = [];
 
 	// If there's a can global
@@ -12,8 +12,18 @@ module.exports = function(doc){
 		});
 	}
 	overrides.push(function(){
+		return new Override(global.canSsr, "request", function(){
+			return req;
+		});
+	});
+	var globalDoc = global.document;
+	overrides.push(function(){
+		return new Override(global.canSsr, "globalDocument", function(){
+			return globalDoc;
+		});
+	});
+	overrides.push(function(){
 		return new Override(global, "document", function(){
-			console.log( "99999999999" );
 			return doc;
 		});
 	});
