+++ lib/middleware/index.js
@@ -6,6 +6,8 @@ var os = require('os');
 var ssr = require('./../index.js');
 var doctype = '<!DOCTYPE html>';
 
+require( "./xhr" )( global );
+
 module.exports = function(system) {
 	var pkgPath = path.join(path.dirname(system.config), 'package.json');
 	var pkg = require(pkgPath);
