+++ lib/middleware/xhr.js
@@ -2,11 +2,7 @@ var url = require("url");
 var xhr = require('xmlhttprequest').XMLHttpRequest;
 var fullUrl = /^https?:\/\//i;
 
-module.exports = function( global, doc,  req ) {
-	var baseUri = req.url || "";
-	if ( req.protocol && req.get ) {
-		baseUri = req.protocol + '://' + req.get( "host" ) + baseUri;
-	}
+module.exports = function( global ) {
 
 	var XHR = global.XMLHttpRequest = function() {
 		xhr.apply(this, arguments);
@@ -15,6 +11,11 @@ module.exports = function( global, doc,  req ) {
 
 		var oldOpen = this.open;
 		this.open = function() {
+			var req = global.canSsr.request;
+			var baseUri = req.url || "";
+			if ( req.protocol && req.get ) {
+				baseUri = req.protocol + '://' + req.get( "host" ) + baseUri;
+			}
 			var args = Array.prototype.slice.call(arguments);
 			var reqURL = args[1];
 
@@ -30,6 +31,7 @@ module.exports = function( global, doc,  req ) {
 	};
 
 	XHR.prototype.send = function () {
+		var req = global.canSsr.request;
 		var cookie = req.headers && req.headers.cookie || "";
 
 		// TODO:
@@ -40,7 +42,7 @@ module.exports = function( global, doc,  req ) {
 			this.setDisableHeaderCheck( true );
 			this.setRequestHeader( "cookie", cookie );
 		}
-		
+
 		return this._hackSend.apply(this, arguments);
 	};
