+++ package.json
@@ -9,11 +9,10 @@
   "scripts": {
     "jshint": "jshint lib/.  test/*.js --config",
     "test:only": "grunt && npm run test:node && npm run test:browser",
+    "test": "npm run jshint && npm run test:only",
 
-    "test:xhrtemporary": "mocha test/cookie_test.js --timeout 60000",
-    "test": "npm run test:xhrtemporary",
-    "originalValueOf_test": "npm run jshint && npm run test:only",
-
+    "justmytest": "mocha --timeout 60000 test/cookie_test.js",
+    
     "test:node": "mocha --timeout 10000 test/test.js && mocha --timeout 10000 test/test_envs.js && mocha test/async_test.js --timeout 10000 && mocha test/jquery_test.js --timeout 10000 && mocha test/react_test.js --timeout 10000 && mocha test/plugin_test.js --timeout 10000 && mocha test/plainjs_test.js --timeout 10000 && mocha test/server_test.js --timeout 10000 && mocha test/middleware_test.js --timeout 10000 && mocha test/can-serve_test.js --timeout 60000 && mocha test/leakscope_test.js --timeout 60000 && mocha test/xhr_test.js --timeout 60000 && mocha test/cookie_test.js --timeout 60000",
     "test:browser": "testee test/test.html --browsers firefox --reporter Spec",
     "publish": "git push origin && git push origin --tags",
