+++ test/tests/cookie/appstate.js
@@ -10,7 +10,6 @@ module.exports = AppMap.extend({
 				var xhr = new XMLHttpRequest();
 
 				xhr.addEventListener("load", function() {
-					console.log( "load happened, set will happen" );
 					set( this.responseText );
 				});
 				xhr.addEventListener("error", function() {
@@ -23,8 +22,6 @@ module.exports = AppMap.extend({
 		},
 		cookie: {
 			get: function ( last, set ) {
-				console.log( "__id:", document.__id ); // why is this undefined??
-				console.log( document.cookie ); // why is this undefined??
 				set( document.cookie );
 			}
 		}
